chore: added view models and documented existing models
