+++ Models/VolunteerViewModels/VolunteerPastEventsViewModel.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+
+namespace Community.Models.VolunteerViewModels
+{
+   /**
+    * Class: VolunteerPastEventsViewModel
+    * Purpose: Contains information about the past events for a user
+    */
+    public class VolunteerPastEventsViewModel
+    {
+        public List<Event> PastEvents { get; set; }
+    }
+}
